Cleaning up a few things and parsing a bit of the data
