+++ BluetoothLEPair/BluetoothLEPair/BluetoothLEPair.csproj
@@ -96,13 +96,19 @@
     <None Include="project.json" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="App.xaml.cs">
+    <Compile Include="Application.cs">
       <DependentUpon>App.xaml</DependentUpon>
     </Compile>
-    <Compile Include="MainPage.xaml.cs">
+    <Compile Include="Bluetooth\BluetoothDataType.cs" />
+    <Compile Include="ViewModel\BluetoothAdvertismentViewerViewModel.cs" />
+    <Compile Include="Bluetooth\BluetoothLookup.cs" />
+    <Compile Include="ViewModel\BluetoothLEAdvertismentViewModel.cs" />
+    <Compile Include="MainPage.cs">
       <DependentUpon>MainPage.xaml</DependentUpon>
     </Compile>
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Input\SimpleCommand.cs" />
+    <Compile Include="ViewModel\NotifyType.cs" />
   </ItemGroup>
   <ItemGroup>
     <AppxManifest Include="Package.appxmanifest">
@@ -130,6 +136,7 @@
       <SubType>Designer</SubType>
     </Page>
   </ItemGroup>
+  <ItemGroup />
   <PropertyGroup Condition=" '$(VisualStudioVersion)' == '' or '$(VisualStudioVersion)' &lt; '14.0' ">
     <VisualStudioVersion>14.0</VisualStudioVersion>
   </PropertyGroup>
