+++ BluetoothLEPair/BluetoothLEPair/MainPage.cs
@@ -1,0 +1,91 @@
+﻿
+
+using System;
+using Windows.UI.Xaml;
+using Windows.UI.Xaml.Controls;
+using Windows.UI.Xaml.Media;
+using Windows.UI.Xaml.Navigation;
+using BluetoothLEPair.ViewModel;
+
+namespace BluetoothLEPair
+{
+    /// <summary>
+    /// An empty page that can be used on its own or navigated to within a Frame.
+    /// </summary>
+    public sealed partial class MainPage : Page
+    {
+        public MainPage()
+        {
+            InitializeComponent();
+        }
+
+        /// <summary>
+        /// Invoked when this page is about to be displayed in a Frame.
+        ///
+        /// We will enable/disable parts of the UI if the device doesn't support it.
+        /// </summary>
+        /// <param name="eventArgs">Event data that describes how this page was reached. The Parameter
+        /// property is typically used to configure the page.</param>
+        protected override void OnNavigatedTo(NavigationEventArgs e)
+        {
+            NotifyUser("Press Run to start watcher.", NotifyType.StatusMessage);
+        }
+
+        /// <summary>
+        /// Invoked immediately before the Page is unloaded and is no longer the current source of a parent Frame.
+        /// </summary>
+        /// <param name="e">
+        /// Event data that can be examined by overriding code. The event data is representative
+        /// of the navigation that will unload the current Page unless canceled. The
+        /// navigation can potentially be canceled by setting Cancel.
+        /// </param>
+        protected override void OnNavigatingFrom(NavigatingCancelEventArgs e)
+        {
+            NotifyUser("Navigating away. Watcher stopped.", NotifyType.StatusMessage);
+            base.OnNavigatingFrom(e);
+        }
+
+        /// <summary>
+        /// Invoked when application execution is being suspended.  Application state is saved
+        /// without knowing whether the application will be terminated or resumed with the contents
+        /// of memory still intact.
+        /// </summary>
+        /// <param name="sender">The source of the suspend request.</param>
+        /// <param name="e">Details about the suspend request.</param>
+        private void App_Suspending(object sender, Windows.ApplicationModel.SuspendingEventArgs e)
+        {
+            NotifyUser("App suspending. Watcher stopped.", NotifyType.StatusMessage);
+        }
+
+        /// Used to display messages to the user
+        /// </summary>
+        /// <param name="strMessage"></param>
+        /// <param name="type"></param>
+        public void NotifyUser(string strMessage, NotifyType type)
+        {
+            switch (type)
+            {
+                case NotifyType.StatusMessage:
+                    ErrorBorder.Background = new SolidColorBrush(Windows.UI.Colors.Green);
+                    break;
+                case NotifyType.ErrorMessage:
+                    ErrorBorder.Background = new SolidColorBrush(Windows.UI.Colors.Red);
+                    break;
+            }
+            StatusBlock.Text = strMessage;
+
+            // Collapse the StatusBlock if it has no text to conserve real estate.
+            ErrorBorder.Visibility = (StatusBlock.Text != String.Empty) ? Visibility.Visible : Visibility.Collapsed;
+            if (StatusBlock.Text != String.Empty)
+            {
+                ErrorBorder.Visibility = Visibility.Visible;
+                StatusBlock.Visibility = Visibility.Visible;
+            }
+            else
+            {
+                ErrorBorder.Visibility = Visibility.Collapsed;
+                StatusBlock.Visibility = Visibility.Collapsed;
+            }
+        }
+    }
+}
