+++ BluetoothLEPair/BluetoothLEPair/MainPage.xaml
@@ -5,65 +5,70 @@
     xmlns:local="using:BluetoothLEPair"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    xmlns:viewModel="using:BluetoothLEPair.ViewModel"
     mc:Ignorable="d"
     x:Name="BLEPairApp">
     <Page.DataContext>
-        <local:BleViewModel />
+        <viewModel:BluetoothAdvertismentViewerViewModel />
     </Page.DataContext>
 
     <Grid Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
-        <Grid x:Name="RootGrid" Margin="12,20,12,12">
+        <Grid>
             <Grid.RowDefinitions>
                 <RowDefinition Height="Auto"/>
                 <RowDefinition Height="*"/>
                 <RowDefinition Height="Auto"/>
             </Grid.RowDefinitions>
-            <StackPanel Margin="0,0,0,10">
+            <StackPanel Grid.Row="0">
                 <TextBlock x:Name="IntroText" TextWrapping="Wrap">
                 </TextBlock>
             </StackPanel>
 
-            <ScrollViewer Grid.Row="1" VerticalScrollMode="Auto" VerticalScrollBarVisibility="Auto">
-                <StackPanel HorizontalAlignment="Left" VerticalAlignment="Top">
-                    <Grid>
-                        <Grid.ColumnDefinitions>
-                            <ColumnDefinition Width="*" />
-                            <ColumnDefinition Width="*" />
-                        </Grid.ColumnDefinitions>
-                        <Grid.RowDefinitions>
-                            <RowDefinition Height="Auto" />
-                            <RowDefinition Height="*" />
-                        </Grid.RowDefinitions>
+            <Grid Grid.Row="1">
+                <Grid.RowDefinitions>
+                    <RowDefinition Height="75" />
+                    <RowDefinition Height="*"/>
+                </Grid.RowDefinitions>
+                <Grid.ColumnDefinitions>
+                    <ColumnDefinition Width="*" />
+                    <ColumnDefinition Width="*" />
+                </Grid.ColumnDefinitions>
 
-                        <Button x:Name="RunButton" Grid.Row="0" Grid.Column="0" Content="Run" Margin="5,10,10,0" Command="{Binding RunCommand}"/>
-                        <Button x:Name="StopButton" Grid.Row="0" Content="Stop" Margin="73,10,0,0" Command="{Binding StopCommand}" />
-                        <Button x:Name="PairButton" Grid.Row="0" Grid.Column="1" Content="Pair" Margin="113,10,0,0" />
+                <Button Grid.Row="0" Grid.Column="0" HorizontalAlignment="Stretch" VerticalAlignment="Stretch"  Margin="3"  Command="{Binding RunCommand}">
+                    <Button.Content>
+                        <Viewbox>
+                            <TextBlock Text="Run" />
+                        </Viewbox>
+                    </Button.Content>
+                </Button>
+                <Button Grid.Row="0" Grid.Column="1" HorizontalAlignment="Stretch" VerticalAlignment="Stretch" Margin="3" Command="{Binding StopCommand}" >
+                    <Button.Content>
+                        <Viewbox>
+                            <TextBlock Text="Stop" />
+                        </Viewbox>
+                    </Button.Content>
+                </Button>
 
-                        <ListBox
-                            x:Name="ReceivedAdvertisementListBox"
+                <ListBox
                             Grid.Row="1"
+                            Grid.Column="0"
                             Grid.ColumnSpan="2"
                             HorizontalAlignment="Stretch"
-                            VerticalAlignment="Top"
-                            Height="300"
-                            Margin="0,10,0,0"
+                            VerticalAlignment="Stretch"
                             ScrollViewer.HorizontalScrollBarVisibility="Auto"
                             ScrollViewer.HorizontalScrollMode="Auto"
                             ItemsSource="{Binding Advertisements.View}"
                            >
-                            <ListBox.ItemTemplate>
-                                <DataTemplate>
-                                    <StackPanel Orientation="Horizontal">
-                                        <TextBlock Text="{Binding manufacturer}" />
-                                        <TextBlock Text="{Binding address}" />
-                                    </StackPanel>
-                                </DataTemplate>
-                            </ListBox.ItemTemplate>
-                        </ListBox>
+                    <ListBox.ItemTemplate>
+                        <DataTemplate>
+                            <StackPanel Orientation="Horizontal">
+                                <TextBlock Text="{Binding Description}" />
+                            </StackPanel>
+                        </DataTemplate>
+                    </ListBox.ItemTemplate>
+                </ListBox>
+            </Grid>
 
-                    </Grid>
-                </StackPanel>
-            </ScrollViewer>
 
             <!-- Status Block for providing messages to the user.  Use the
              NotifyUser() method to populate the message -->
