+++ BluetoothLEPair/BluetoothLEPair/ViewModel/NotifyType.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace BluetoothLEPair.ViewModel
+{
+    public enum NotifyType
+    {
+        StatusMessage,
+        ErrorMessage
+    }
+}
